+++ dump_tool/src/EvidenceBuilderInternalsPriv.h
@@ -64,6 +64,7 @@ std::wstring Hex64(std::uint64_t v);
 
 std::optional<std::wstring> TryExplainExceptionInfo(const AnalysisResult& r, bool en);
 bool IsSystemishModule(std::wstring_view filename);
+bool IsLikelyWindowsSystemModulePath(std::wstring_view modulePath);
 bool IsGameExeModule(std::wstring_view filename);
 bool IsKnownHookFramework(std::wstring_view filename);